% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmse.R
\name{get_rmse}
\alias{get_rmse}
\title{Get RMSE for two matrices}
\usage{
get_rmse(ref.q, test.q)
}
\arguments{
\item{ref.q}{\link{matrix} Numeric values of N x K}

\item{test.q}{\link{matrix} Numeric values of N x K}
}
\value{
Numeric value
}
\description{
Get RMSE for two matrices
}
\examples{
test.1 <- matrix(1, nrow = 2, ncol = 2)
test.2 <- matrix(2, nrow = 2, ncol = 2)
get_rmse(test.1, test.2)
}
